refactor: remove useBestProducts and create useProductsPagination for unified pagination and sort logic
