+++ src/pages/Items/components/AllProducts/AllProducts.js
@@ -1,22 +1,11 @@
 import styles from "./AllProducts.module.css";
-import useResponsiveLimit from "../../hooks/useResponsiveLimit";
-import usePaginationState from "../../hooks/usePaginationState";
-import usePaginatedProducts from "../../hooks/usePaginatedProducts";
+import useProductsPagination from "../../hooks/useProductsPagination";
 import ProductSection from "../ProductSection/ProductSection";
 import Pagination from "../Pagination/Pagination";
-import { useState, useCallback } from "react";
 
 function AllProducts({ title, itemsPerDevice }) {
-  const limit = useResponsiveLimit(itemsPerDevice);
-  const [page, changePage] = usePaginationState(limit);
-  const [sort, setSort] = useState("latest");
-
-  const { products, totalPages } = usePaginatedProducts({ page, limit, sort });
-
-  const handleSortChange = useCallback((newSort) => {
-    setSort(newSort);
-    changePage(1);
-  }, [changePage]);
+  const { products, totalPages, page, changePage, sort, handleSortChange, } =
+    useProductsPagination(itemsPerDevice);
 
   return (
     <>
@@ -29,15 +18,14 @@ function AllProducts({ title, itemsPerDevice }) {
         onChangeSort={handleSortChange}
       />
       <div className={styles.paginationWrapper}>
-      <Pagination
-        currentPage={page}
-        totalPages={totalPages}
-        onPageChange={changePage}
-      />
+        <Pagination
+          currentPage={page}
+          totalPages={totalPages}
+          onPageChange={changePage}
+        />
       </div>
     </>
   );
 }
-
 
 export default AllProducts;
