+++ src/pages/Items/components/BestProducts/BestProducts.js
@@ -1,11 +1,31 @@
-import useBestProducts from "../../hooks/useBestProducts";
+import { useEffect, useState } from "react";
+import { fetchProducts } from "../../../../api/products";
+import { getLimitFromWindowWidth } from "../../../../utils/getLimitFromWindowWidth";
 import ProductSection from "../ProductSection/ProductSection";
 
-function BestProducts({ title, itemsPerDevice}) {
- 
-  const bestProducts = useBestProducts(itemsPerDevice);
+function BestProducts({ title, itemsPerDevice }) {
+  const [bestProducts, setBestProducts] = useState([]);
 
-return <ProductSection title={title} products={bestProducts}/>;
+  useEffect(() => {
+    const updateProducts = async () => {
+      const data = await fetchProducts();
+      const sorted = data.sort((a, b) => b.favoriteCount - a.favoriteCount);
+      const limit = getLimitFromWindowWidth(
+        itemsPerDevice.desktop,
+        itemsPerDevice.tablet,
+        itemsPerDevice.mobile
+      );
+      setBestProducts(sorted.slice(0, limit));
+    };
+
+    updateProducts();
+    const resizeHandler = () => updateProducts();
+    window.addEventListener("resize", resizeHandler);
+
+    return () => window.removeEventListener("resize", resizeHandler);
+  }, [itemsPerDevice]);
+
+  return <ProductSection title={title} products={bestProducts} />;
 }
 
 export default BestProducts;
