--- src/pages/Items/hooks/useBestProducts.js
@@ -1,27 +1,0 @@
-import { useEffect, useState } from "react";
-import { fetchProducts } from "../../../api/products";
-import { getLimitFromWindowWidth } from "../../../utils/getLimitFromWindowWidth";
-
-export default function useBestProducts(itemsPerDevice) {
-  const [bestProducts, setBestProducts] = useState([]);
-
-  useEffect(() => {
-    const updateProducts = async () => {
-      const data = await fetchProducts();
-      const sorted = data.sort((a, b) => b.favoriteCount - a.favoriteCount);
-      const limit = getLimitFromWindowWidth(
-        itemsPerDevice.desktop,
-        itemsPerDevice.tablet,
-        itemsPerDevice.mobile
-      );
-      setBestProducts(sorted.slice(0, limit));
-    };
-
-    updateProducts();
-    const resizeHandler = () => updateProducts();
-    window.addEventListener("resize", resizeHandler);
-
-    return () => window.removeEventListener("resize", resizeHandler);
-  }, [itemsPerDevice]);
-  return bestProducts;
-}
